change db arch

+++ crates/tml_infrastructure/src/entity/app/music_info.rs
@@ -13,8 +13,8 @@ pub struct Model {
     pub track_number: i32,
     pub audio_bitrate: i32,
     pub sample_rate: i32,
-    pub channels: i8,
-    pub bit_depth: i8,
+    pub channels: i16,
+    pub bit_depth: i16,
     pub storage_id: i64,
     pub file_path: String,
     #[sea_orm(has_many, via = "music_info_music_list")]
